requirements.txt
Django==1.5.1
MySQL-python==1.2.4
argparse==1.2.1
djangorestframework==2.3.2
wsgiref==0.1.2
